+++ ArctangentFunction.cs
@@ -29,7 +29,7 @@ namespace MetaphysicsIndustries.Solus
 
         protected override Literal InternalCall(VariableTable varTable, Literal[] args)
 		{
-			return new Literal(Math.Atan(args[0].Eval(varTable).Value));
+            return new Literal((float)Math.Atan(args[0].Eval(varTable).Value));
 		}
 
         public override string DisplayName
