+++ LessThanComparisonOperation.cs
@@ -11,7 +11,7 @@ namespace MetaphysicsIndustries.Solus
         {
         }
 
-        protected override bool Compare(double x, double y)
+        protected override bool Compare(float x, float y)
         {
             return x < y;
         }
