+++ SolusTensor.cs
@@ -13,8 +13,8 @@ namespace MetaphysicsIndustries.Solus
             return GetEnumerator();
         }
 
-        public delegate double Modulator(double x);
-        public delegate double BiModulator(double x, double y);
+        public delegate float Modulator(float x);
+        public delegate float BiModulator(float x, float y);
         public abstract void ApplyToAll(Modulator mod);
     }
 }
